adds release pipeline flows and android builds
- release pipeline
- renames org.dif to org.didcommx
- restructures project
- adds android testing
