+++ build.gradle
@@ -1,49 +1,21 @@
-plugins {
-    id 'java-library'
-    id 'org.jetbrains.kotlin.jvm' version "${kotlinJvm}"
-    id "org.jlleitschuh.gradle.ktlint" version "${ktlintGradle}"
-    // TODO: remove once published to Maven
-    id 'com.github.johnrengelman.shadow' version '7.0.0'
-}
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+buildscript {
+    repositories {
+        google()
+        mavenCentral()
+    }
 
-group = 'org.dif.peerdid'
-version = '0.1-SNAPSHOT'
+    dependencies {
+        classpath "com.android.tools.build:gradle:${androidGradle}"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${kotlinJvm}"
 
-repositories {
-    mavenCentral()
-    maven {
-        url 'https://maven.scijava.org/content/groups/public'
+        // jUnit 5 support
+        classpath("de.mannodermaus.gradle.plugins:android-junit5:${mannodermausJUnit5}")
+
+        // classpath 'androidx.benchmark:benchmark-gradle-plugin:1.0.0'
     }
 }
 
-dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib:1.5.10'
-    // https://mvnrepository.com/artifact/com.google.code.gson/gson
-    implementation 'com.google.code.gson:gson:2.8.8'
-    // https://mvnrepository.com/artifact/com.github.multiformats/java-multibase
-    implementation 'com.github.multiformats:java-multibase:v1.1.0'
-    implementation "com.zmannotes:varint:1.0.0"
-    testImplementation 'org.jetbrains.kotlin:kotlin-test'
-    testImplementation 'org.junit.jupiter:junit-jupiter-params:5.7.0'
-}
-
-test {
-    useJUnitPlatform()
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "${jvmTarget}"
-}
-
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "${jvmTarget}"
-}
-
-
-shadowJar {
-    // @see https://youtrack.jetbrains.com/issue/KT-25709
-    exclude("**/*.kotlin_metadata")
-    exclude("**/*.kotlin_builtins")
-
-    archiveClassifier.set("") // remove suffix `-all` as intellij can't find the library otherwise
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
