+++ gradle.properties
@@ -1,4 +1,43 @@
+# Project-wide Gradle settings.
+
+version = '0.1.0-SNAPSHOT'
+
+# Kotlin code style for this project: "official" or "obsolete":
+kotlin.code.style=official
+
 kotlin.code.style=official
 kotlinJvm=1.5.21
 jvmTarget=1.8
 ktlintGradle=10.1.0
+
+androidGradle=7.0.0
+mannodermausJUnit5=1.8.0.0
+androidxTestRunner=1.4.0
+jUnitJupiter=5.8.0
+mannodermausJUunit5AndroidTest=1.3.0
+googleGson=2.8.8
+javaMultibase=v1.1.0
+johnrengelmanShadow=7.0.0
+zmanVarint=1.0.0
+
+# Android settigns
+# androidBuilds=true
+
+# IDE (e.g. Android Studio) users:
+# Gradle settings configured through the IDE *will override*
+# any settings specified in this file.
+# For more details on how to configure your build environment visit
+# http://www.gradle.org/docs/current/userguide/build_environment.html
+# Specifies the JVM arguments used for the daemon process.
+# The setting is particularly useful for tweaking memory settings.
+org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
+# When configured, Gradle will run in incubating parallel mode.
+# This option should only be used with decoupled projects. More details, visit
+# http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
+# org.gradle.parallel=true
+# AndroidX package structure to make it clearer which packages are bundled with the
+# Android operating system, and which are packaged with your app"s APK
+# https://developer.android.com/topic/libraries/support-library/androidx-rn
+android.useAndroidX=true
+# Automatically convert third-party libraries to use AndroidX
+android.enableJetifier=true
